Show progress bars during image uploading. Fixes #34.
This is a patch for the problem of our progress bars not showing during
image upload. The problem happened to be a nesting issue with the DOM
element that holds the actual bar. It needed to be nested inside of
another element that, by accident, was appearing as a sibling instead of
a parent.

Signed-off-by: David Celis <david@davidcelis.com>