Rework the models for Actions and Attributes
This commit addresses an issue with how Actions and Attributes were
defined within the Rails application. There were several design flaws
that have been taken care of here:

1. We now have ActionDefinitions and AttributeDefinitions. These are
models for actions we have defined and their defined attributes. This
table will be seeded by us, and these will be the actual CCActions for
cocos2d.

2. Actions and Attributes are now instances of defined actions and
defined attributes which are directly associated to scenes based on what
a user would like their scene to do.

  * Actions are hooked up to a Scene and an ActionDefinition.
  * Attributes are hooked up to an Action, a Keyframe, and an
  AttributeDefinition. This is where the user will supply an actual value
  for a specified attribute.

3. The ActionGroup model and join table has been entirely removed. I saw
this as unnecessary now that an Action can be associated directly to a
Scene.

Signed-off-by: David Celis <david@davidcelis.com>