Refactoring
Note: I don't know how to test this code, so I
refactored it 'blindly'. Might have broken things.

"Feature Envy: Methods and code parts that
make extensive use of another class may belong in
the other class. Consider moving this method to 
the class it is so envious of."

Also, triggering an event on an object (keyframe,
in this case) and having that object listening on
itself for that event is not a good practice. It's
better just to invoke a method on the object and
let it to everything.

As a general rule, think about introducing new events.
Are they really necessary, used for decoupling
different things?