Merge branch 'refactor-widgets' into refactor
I commented out a lot of code that depends on the widget system.
The idea is to get the basic widgets infrastructure in place,
with the ability to switch between scenes, keyframes, and have the
widgets rendered properly. Afterwards we'll re-enable the other
functions.

Please see
https://github.com/curiousminds/interapptive-web/wiki/Refactoring-widgets

This will probably break a lot of stuff. Appologies! and fix it if you
can :D

Conflicts:
	app/assets/javascripts/app.js.coffee
	app/assets/javascripts/app/models/image.js.coffee
	app/assets/javascripts/app/models/keyframe.js.coffee
	app/assets/javascripts/app/models/scene.js.coffee
	app/assets/javascripts/app/templates/assets/library.jst.hamlc
	app/assets/javascripts/app/views/assets/images/index.js.coffee
	app/assets/javascripts/app/views/assets/library.js.coffee
	app/assets/javascripts/app/views/keyframes/index.js.coffee
	app/assets/javascripts/app/views/toolbar.js.coffee
	app/assets/javascripts/builder/widgets/sprite_widget.js.coffee
	app/assets/javascripts/builder/widgets/text_widget.js.coffee