Fixed multiple asset upload and deletion #469
It did not work anymore after we changed the view
to add/remove to the Backbone collection (and serenader)
instead of fetching.
Adding multiple assets resulted in a sequence of
calls to `render`. Which re-initialized the upload
plugin and lost the old handlers.

* do not re-initialize the upload plugin when
the data changes; just re-render the assets
* let the upload plugin handle all the UI; do not
re-render when the upload plugin adds assets to the
collection
* removed the advanced table plugin. It interferes
with jqueryupload, messing its rendering and
expectations