Save scene once when multiple changes are made
at the save time.

The zOrder bug happened because, when the top
sprite was moved to the bottom, the zOrder of each
widget changed, causing the scene to be saved
multiple times (the same number of times as the
number of scene widgets).
(On development, the requests are received and
executed in the order they are issued, so the bug
does not occur.)