ISSUE #555 Remove corresponding widgets if an asset is removed from asset library.
ISSUE #555 Destroy text widgets on destruction of a font only when there
is no default font available. In case there is a default font available,
set font_id in text widgets to default font's id.

ISSUE #555 Get sound as an Sound object for scene background sound. As a
byproduct synchronize scene's background sound when it's deleted.