(#662) Context menu for main menu buttons.
(#662) It does not make sense to let users put a button widget to back
or in front of other sprites. Those should always be on top.

Remove the active images palette #662

Bring to front and put in back - in the model layer #662

Remove everything related to palettes #662

(#662) Provide coordinate and scale related options for button widgets.