(#992) Do not use same classname for differentiating between highlight state and actually highlighting a word.
When audio is done playing, popcorn removes class 'highlighted' from the
words. This sometimes errors in letting a word to be ready for
highlighting even though the word before it is not highlighted and it's
not the first word. Using a separate class to indicate processing state
of a word works fine.